backups/lexigraph-12_7-20250819-170413/requirements.txt
streamlit
pymupdf
pillow
easyocr
scikit-learn
rdflib
joblib
